# internal/controller/handler/v1/purchases/detail/ship/gen/server.gen.go
// Package gen provides primitives to interact with the openapi HTTP API.
//
// Code generated by github.com/oapi-codegen/oapi-codegen/v2 version v2.8.0 DO NOT EDIT.
package gen

import (
	"bytes"
	"context"
	"encoding/json"
	"fmt"
	"net/http"

	"github.com/labstack/echo/v5"
	"github.com/oapi-codegen/runtime"
)

// ServerInterface represents all server handlers.
type ServerInterface interface {
	// PatchPurchasesShip 購入の発送
	// (PATCH /v1/purchases/{purchaseCode}/ship)
	PatchPurchasesShip(ctx *echo.Context, purchaseCode PurchaseCodeParam) error
}

// ServerInterfaceWrapper converts echo contexts to parameters.
type ServerInterfaceWrapper struct {
	Handler ServerInterface
}

// PatchPurchasesShip converts echo context to params.
func (w *ServerInterfaceWrapper) PatchPurchasesShip(ctx *echo.Context) error {
	var err error
	// ------------- Path parameter "purchaseCode" -------------
	var purchaseCode PurchaseCodeParam

	err = runtime.BindStyledParameterWithOptions("simple", "purchaseCode", ctx.Param("purchaseCode"), &purchaseCode, runtime.BindStyledParameterOptions{ParamLocation: runtime.ParamLocationPath, Explode: false, Required: true, Type: "string", Format: "", ValueIsUnescaped: ctx.Request().URL.RawPath == ""})
	if err != nil {
		return echo.NewHTTPError(http.StatusBadRequest, fmt.Sprintf("Invalid format for parameter purchaseCode: %s", err))
	}

	// Invoke the callback with all the unmarshaled arguments
	err = w.Handler.PatchPurchasesShip(ctx, purchaseCode)
	return err
}

// This is a simple interface which specifies echo.Route addition functions which
// are present on both echo.Echo and echo.Group, since we want to allow using
// either of them for path registration
type EchoRouter interface {
	CONNECT(path string, h echo.HandlerFunc, m ...echo.MiddlewareFunc) echo.RouteInfo
	DELETE(path string, h echo.HandlerFunc, m ...echo.MiddlewareFunc) echo.RouteInfo
	GET(path string, h echo.HandlerFunc, m ...echo.MiddlewareFunc) echo.RouteInfo
	HEAD(path string, h echo.HandlerFunc, m ...echo.MiddlewareFunc) echo.RouteInfo
	OPTIONS(path string, h echo.HandlerFunc, m ...echo.MiddlewareFunc) echo.RouteInfo
	PATCH(path string, h echo.HandlerFunc, m ...echo.MiddlewareFunc) echo.RouteInfo
	POST(path string, h echo.HandlerFunc, m ...echo.MiddlewareFunc) echo.RouteInfo
	PUT(path string, h echo.HandlerFunc, m ...echo.MiddlewareFunc) echo.RouteInfo
	TRACE(path string, h echo.HandlerFunc, m ...echo.MiddlewareFunc) echo.RouteInfo
}

// RegisterHandlersOptions configures RegisterHandlersWithOptions.
type RegisterHandlersOptions struct {
	// BaseURL is prepended to every registered path so the API can be served
	// under a prefix.
	BaseURL string
	// OperationMiddlewares lets the caller attach per-operation middleware at
	// registration time. The map key is the OpenAPI `operationId` value as it
	// appears in the spec (the raw, un-normalized form). Operations that have
	// no entry are registered with no extra middleware. A nil map disables
	// per-operation middleware entirely.
	OperationMiddlewares map[string][]echo.MiddlewareFunc
}

// RegisterHandlers adds each server route to the EchoRouter.
func RegisterHandlers(router EchoRouter, si ServerInterface) {
	RegisterHandlersWithOptions(router, si, RegisterHandlersOptions{})
}

// RegisterHandlersWithBaseURL registers handlers and prepends BaseURL to the
// paths so the API can be served under a prefix.
func RegisterHandlersWithBaseURL(router EchoRouter, si ServerInterface, baseURL string) {
	RegisterHandlersWithOptions(router, si, RegisterHandlersOptions{BaseURL: baseURL})
}

// RegisterHandlersWithOptions registers handlers using the supplied options,
// including any per-operation middleware.
func RegisterHandlersWithOptions(router EchoRouter, si ServerInterface, options RegisterHandlersOptions) {

	wrapper := ServerInterfaceWrapper{
		Handler: si,
	}

	router.PATCH(options.BaseURL+"/v1/purchases/:purchaseCode/ship", wrapper.PatchPurchasesShip, options.OperationMiddlewares["PatchPurchasesShip"]...)

}

type BadRequest400JSONResponse ErrorResponse

type Conflict409JSONResponse ErrorResponse

type Forbidden403JSONResponse ErrorResponse

type InternalServerError500JSONResponse ErrorResponse

type MethodNotAllowed405JSONResponse ErrorResponse

type NotFound404JSONResponse ErrorResponse

type Unauthorized401JSONResponse ErrorResponse

type PatchPurchasesShipRequestObject struct {
	PurchaseCode PurchaseCodeParam `json:"purchaseCode"`
}

type PatchPurchasesShipResponseObject interface {
	VisitPatchPurchasesShipResponse(w http.ResponseWriter) error
}

type PatchPurchasesShip200JSONResponse PurchaseShipResponse

func (response PatchPurchasesShip200JSONResponse) VisitPatchPurchasesShipResponse(w http.ResponseWriter) error {

	var buf bytes.Buffer
	if err := json.NewEncoder(&buf).Encode(response); err != nil {
		return err
	}
	w.Header().Set("Content-Type", "application/json")
	w.WriteHeader(200)
	_, err := buf.WriteTo(w)
	return err
}

type PatchPurchasesShip400JSONResponse struct{ BadRequest400JSONResponse }

func (response PatchPurchasesShip400JSONResponse) VisitPatchPurchasesShipResponse(w http.ResponseWriter) error {

	var buf bytes.Buffer
	if err := json.NewEncoder(&buf).Encode(response); err != nil {
		return err
	}
	w.Header().Set("Content-Type", "application/json")
	w.WriteHeader(400)
	_, err := buf.WriteTo(w)
	return err
}

type PatchPurchasesShip401JSONResponse struct{ Unauthorized401JSONResponse }

func (response PatchPurchasesShip401JSONResponse) VisitPatchPurchasesShipResponse(w http.ResponseWriter) error {

	var buf bytes.Buffer
	if err := json.NewEncoder(&buf).Encode(response); err != nil {
		return err
	}
	w.Header().Set("Content-Type", "application/json")
	w.WriteHeader(401)
	_, err := buf.WriteTo(w)
	return err
}

type PatchPurchasesShip403JSONResponse struct{ Forbidden403JSONResponse }

func (response PatchPurchasesShip403JSONResponse) VisitPatchPurchasesShipResponse(w http.ResponseWriter) error {

	var buf bytes.Buffer
	if err := json.NewEncoder(&buf).Encode(response); err != nil {
		return err
	}
	w.Header().Set("Content-Type", "application/json")
	w.WriteHeader(403)
	_, err := buf.WriteTo(w)
	return err
}

type PatchPurchasesShip404JSONResponse struct{ NotFound404JSONResponse }

func (response PatchPurchasesShip404JSONResponse) VisitPatchPurchasesShipResponse(w http.ResponseWriter) error {

	var buf bytes.Buffer
	if err := json.NewEncoder(&buf).Encode(response); err != nil {
		return err
	}
	w.Header().Set("Content-Type", "application/json")
	w.WriteHeader(404)
	_, err := buf.WriteTo(w)
	return err
}

type PatchPurchasesShip405JSONResponse struct {
	MethodNotAllowed405JSONResponse
}

func (response PatchPurchasesShip405JSONResponse) VisitPatchPurchasesShipResponse(w http.ResponseWriter) error {

	var buf bytes.Buffer
	if err := json.NewEncoder(&buf).Encode(response); err != nil {
		return err
	}
	w.Header().Set("Content-Type", "application/json")
	w.WriteHeader(405)
	_, err := buf.WriteTo(w)
	return err
}

type PatchPurchasesShip409JSONResponse struct{ Conflict409JSONResponse }

func (response PatchPurchasesShip409JSONResponse) VisitPatchPurchasesShipResponse(w http.ResponseWriter) error {

	var buf bytes.Buffer
	if err := json.NewEncoder(&buf).Encode(response); err != nil {
		return err
	}
	w.Header().Set("Content-Type", "application/json")
	w.WriteHeader(409)
	_, err := buf.WriteTo(w)
	return err
}

type PatchPurchasesShip500JSONResponse struct {
	InternalServerError500JSONResponse
}

func (response PatchPurchasesShip500JSONResponse) VisitPatchPurchasesShipResponse(w http.ResponseWriter) error {

	var buf bytes.Buffer
	if err := json.NewEncoder(&buf).Encode(response); err != nil {
		return err
	}
	w.Header().Set("Content-Type", "application/json")
	w.WriteHeader(500)
	_, err := buf.WriteTo(w)
	return err
}

// StrictServerInterface represents all server handlers.
type StrictServerInterface interface {
	// PatchPurchasesShip 購入の発送
	// (PATCH /v1/purchases/{purchaseCode}/ship)
	PatchPurchasesShip(ctx context.Context, request PatchPurchasesShipRequestObject) (PatchPurchasesShipResponseObject, error)
}

type StrictHandlerFunc func(ctx *echo.Context, request any) (any, error)
type StrictMiddlewareFunc func(f StrictHandlerFunc, operationID string) StrictHandlerFunc

func NewStrictHandler(ssi StrictServerInterface, middlewares []StrictMiddlewareFunc) ServerInterface {
	return &strictHandler{ssi: ssi, middlewares: middlewares}
}

type strictHandler struct {
	ssi         StrictServerInterface
	middlewares []StrictMiddlewareFunc
}

// PatchPurchasesShip operation middleware
func (sh *strictHandler) PatchPurchasesShip(ctx *echo.Context, purchaseCode PurchaseCodeParam) error {
	var request PatchPurchasesShipRequestObject

	request.PurchaseCode = purchaseCode

	handler := func(ctx *echo.Context, request interface{}) (interface{}, error) {
		return sh.ssi.PatchPurchasesShip(ctx.Request().Context(), request.(PatchPurchasesShipRequestObject))
	}
	for _, middleware := range sh.middlewares {
		handler = middleware(handler, "PatchPurchasesShip")
	}

	response, err := handler(ctx, request)

	if err != nil {
		return err
	} else if validResponse, ok := response.(PatchPurchasesShipResponseObject); ok {
		return validResponse.VisitPatchPurchasesShipResponse(ctx.Response())
	} else if response != nil {
		return fmt.Errorf("unexpected response type: %T", response)
	}
	return nil
}