# internal/controller/handler/v1/purchases/shippable/gen/server.gen.go
// Package gen provides primitives to interact with the openapi HTTP API.
//
// Code generated by github.com/oapi-codegen/oapi-codegen/v2 version v2.8.0 DO NOT EDIT.
package gen

import (
	"bytes"
	"context"
	"encoding/json"
	"fmt"
	"net/http"

	"github.com/labstack/echo/v5"
	"github.com/oapi-codegen/runtime"
)

// ServerInterface represents all server handlers.
type ServerInterface interface {
	// GetPurchasesShippable 発送待ち購入のまとめ発送一覧の取得
	// (GET /v1/purchases/shippable)
	GetPurchasesShippable(ctx *echo.Context, params GetPurchasesShippableParams) error
}

// ServerInterfaceWrapper converts echo contexts to parameters.
type ServerInterfaceWrapper struct {
	Handler ServerInterface
}

// GetPurchasesShippable converts echo context to params.
func (w *ServerInterfaceWrapper) GetPurchasesShippable(ctx *echo.Context) error {
	var err error

	// Parameter object where we will unmarshal all parameters from the context
	var params GetPurchasesShippableParams
	// ------------- Optional query parameter "limit" -------------

	err = runtime.BindQueryParameterWithOptions("form", true, false, "limit", ctx.QueryParams(), &params.Limit, runtime.BindQueryParameterOptions{Type: "integer", Format: ""})
	if err != nil {
		return echo.NewHTTPError(http.StatusBadRequest, fmt.Sprintf("Invalid format for parameter limit: %s", err))
	}

	// Invoke the callback with all the unmarshaled arguments
	err = w.Handler.GetPurchasesShippable(ctx, params)
	return err
}

// This is a simple interface which specifies echo.Route addition functions which
// are present on both echo.Echo and echo.Group, since we want to allow using
// either of them for path registration
type EchoRouter interface {
	CONNECT(path string, h echo.HandlerFunc, m ...echo.MiddlewareFunc) echo.RouteInfo
	DELETE(path string, h echo.HandlerFunc, m ...echo.MiddlewareFunc) echo.RouteInfo
	GET(path string, h echo.HandlerFunc, m ...echo.MiddlewareFunc) echo.RouteInfo
	HEAD(path string, h echo.HandlerFunc, m ...echo.MiddlewareFunc) echo.RouteInfo
	OPTIONS(path string, h echo.HandlerFunc, m ...echo.MiddlewareFunc) echo.RouteInfo
	PATCH(path string, h echo.HandlerFunc, m ...echo.MiddlewareFunc) echo.RouteInfo
	POST(path string, h echo.HandlerFunc, m ...echo.MiddlewareFunc) echo.RouteInfo
	PUT(path string, h echo.HandlerFunc, m ...echo.MiddlewareFunc) echo.RouteInfo
	TRACE(path string, h echo.HandlerFunc, m ...echo.MiddlewareFunc) echo.RouteInfo
}

// RegisterHandlersOptions configures RegisterHandlersWithOptions.
type RegisterHandlersOptions struct {
	// BaseURL is prepended to every registered path so the API can be served
	// under a prefix.
	BaseURL string
	// OperationMiddlewares lets the caller attach per-operation middleware at
	// registration time. The map key is the OpenAPI `operationId` value as it
	// appears in the spec (the raw, un-normalized form). Operations that have
	// no entry are registered with no extra middleware. A nil map disables
	// per-operation middleware entirely.
	OperationMiddlewares map[string][]echo.MiddlewareFunc
}

// RegisterHandlers adds each server route to the EchoRouter.
func RegisterHandlers(router EchoRouter, si ServerInterface) {
	RegisterHandlersWithOptions(router, si, RegisterHandlersOptions{})
}

// RegisterHandlersWithBaseURL registers handlers and prepends BaseURL to the
// paths so the API can be served under a prefix.
func RegisterHandlersWithBaseURL(router EchoRouter, si ServerInterface, baseURL string) {
	RegisterHandlersWithOptions(router, si, RegisterHandlersOptions{BaseURL: baseURL})
}

// RegisterHandlersWithOptions registers handlers using the supplied options,
// including any per-operation middleware.
func RegisterHandlersWithOptions(router EchoRouter, si ServerInterface, options RegisterHandlersOptions) {

	wrapper := ServerInterfaceWrapper{
		Handler: si,
	}

	router.GET(options.BaseURL+"/v1/purchases/shippable", wrapper.GetPurchasesShippable, options.OperationMiddlewares["GetPurchasesShippable"]...)

}

type BadRequest400JSONResponse ErrorResponse

type Forbidden403JSONResponse ErrorResponse

type InternalServerError500JSONResponse ErrorResponse

type MethodNotAllowed405JSONResponse ErrorResponse

type ServiceUnavailable503JSONResponse ErrorResponse

type Unauthorized401JSONResponse ErrorResponse

type GetPurchasesShippableRequestObject struct {
	Params GetPurchasesShippableParams
}

type GetPurchasesShippableResponseObject interface {
	VisitGetPurchasesShippableResponse(w http.ResponseWriter) error
}

type GetPurchasesShippable200JSONResponse PurchaseShippableResponse

func (response GetPurchasesShippable200JSONResponse) VisitGetPurchasesShippableResponse(w http.ResponseWriter) error {

	var buf bytes.Buffer
	if err := json.NewEncoder(&buf).Encode(response); err != nil {
		return err
	}
	w.Header().Set("Content-Type", "application/json")
	w.WriteHeader(200)
	_, err := buf.WriteTo(w)
	return err
}

type GetPurchasesShippable400JSONResponse struct{ BadRequest400JSONResponse }

func (response GetPurchasesShippable400JSONResponse) VisitGetPurchasesShippableResponse(w http.ResponseWriter) error {

	var buf bytes.Buffer
	if err := json.NewEncoder(&buf).Encode(response); err != nil {
		return err
	}
	w.Header().Set("Content-Type", "application/json")
	w.WriteHeader(400)
	_, err := buf.WriteTo(w)
	return err
}

type GetPurchasesShippable401JSONResponse struct{ Unauthorized401JSONResponse }

func (response GetPurchasesShippable401JSONResponse) VisitGetPurchasesShippableResponse(w http.ResponseWriter) error {

	var buf bytes.Buffer
	if err := json.NewEncoder(&buf).Encode(response); err != nil {
		return err
	}
	w.Header().Set("Content-Type", "application/json")
	w.WriteHeader(401)
	_, err := buf.WriteTo(w)
	return err
}

type GetPurchasesShippable403JSONResponse struct{ Forbidden403JSONResponse }

func (response GetPurchasesShippable403JSONResponse) VisitGetPurchasesShippableResponse(w http.ResponseWriter) error {

	var buf bytes.Buffer
	if err := json.NewEncoder(&buf).Encode(response); err != nil {
		return err
	}
	w.Header().Set("Content-Type", "application/json")
	w.WriteHeader(403)
	_, err := buf.WriteTo(w)
	return err
}

type GetPurchasesShippable405JSONResponse struct {
	MethodNotAllowed405JSONResponse
}

func (response GetPurchasesShippable405JSONResponse) VisitGetPurchasesShippableResponse(w http.ResponseWriter) error {

	var buf bytes.Buffer
	if err := json.NewEncoder(&buf).Encode(response); err != nil {
		return err
	}
	w.Header().Set("Content-Type", "application/json")
	w.WriteHeader(405)
	_, err := buf.WriteTo(w)
	return err
}

type GetPurchasesShippable500JSONResponse struct {
	InternalServerError500JSONResponse
}

func (response GetPurchasesShippable500JSONResponse) VisitGetPurchasesShippableResponse(w http.ResponseWriter) error {

	var buf bytes.Buffer
	if err := json.NewEncoder(&buf).Encode(response); err != nil {
		return err
	}
	w.Header().Set("Content-Type", "application/json")
	w.WriteHeader(500)
	_, err := buf.WriteTo(w)
	return err
}

type GetPurchasesShippable503JSONResponse struct {
	ServiceUnavailable503JSONResponse
}

func (response GetPurchasesShippable503JSONResponse) VisitGetPurchasesShippableResponse(w http.ResponseWriter) error {

	var buf bytes.Buffer
	if err := json.NewEncoder(&buf).Encode(response); err != nil {
		return err
	}
	w.Header().Set("Content-Type", "application/json")
	w.WriteHeader(503)
	_, err := buf.WriteTo(w)
	return err
}

// StrictServerInterface represents all server handlers.
type StrictServerInterface interface {
	// GetPurchasesShippable 発送待ち購入のまとめ発送一覧の取得
	// (GET /v1/purchases/shippable)
	GetPurchasesShippable(ctx context.Context, request GetPurchasesShippableRequestObject) (GetPurchasesShippableResponseObject, error)
}

type StrictHandlerFunc func(ctx *echo.Context, request any) (any, error)
type StrictMiddlewareFunc func(f StrictHandlerFunc, operationID string) StrictHandlerFunc

func NewStrictHandler(ssi StrictServerInterface, middlewares []StrictMiddlewareFunc) ServerInterface {
	return &strictHandler{ssi: ssi, middlewares: middlewares}
}

type strictHandler struct {
	ssi         StrictServerInterface
	middlewares []StrictMiddlewareFunc
}

// GetPurchasesShippable operation middleware
func (sh *strictHandler) GetPurchasesShippable(ctx *echo.Context, params GetPurchasesShippableParams) error {
	var request GetPurchasesShippableRequestObject

	request.Params = params

	handler := func(ctx *echo.Context, request interface{}) (interface{}, error) {
		return sh.ssi.GetPurchasesShippable(ctx.Request().Context(), request.(GetPurchasesShippableRequestObject))
	}
	for _, middleware := range sh.middlewares {
		handler = middleware(handler, "GetPurchasesShippable")
	}

	response, err := handler(ctx, request)

	if err != nil {
		return err
	} else if validResponse, ok := response.(GetPurchasesShippableResponseObject); ok {
		return validResponse.VisitGetPurchasesShippableResponse(ctx.Response())
	} else if response != nil {
		return fmt.Errorf("unexpected response type: %T", response)
	}
	return nil
}