# openapi/openapi.gen.yaml
openapi: 3.0.3
info:
  title: Go Boilerplate API
  termsOfService: https://github.com/Tomy-ch/go-boilerplate-echo-oapi-sqlc
  contact:
    name: TomyC258460
    url: https://x.com/TomyC258460
  version: 2.2.0
  description: |
    Reference specification for the sample API (users, products, purchases). It exercises every layer
    of the Onion Architecture end to end: cursor and offset pagination, creation and partial update,
    soft delete, idempotency keys, and role-based authorization.

    Handlers (oapi-codegen) and the published reference documentation are both generated from this
    file, so every endpoint change starts here.
  license:
    name: MIT License
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:8080
    description: Local development server
    x-internal: true
  - url: https://api.example.com
    description: Production server (replace with actual URL)
    x-internal: false
  - url: https://staging.api.example.com
    description: Staging server (replace with actual URL)
    x-internal: false
  - url: https://development.api.example.com
    description: Development server (replace with actual URL)
    x-internal: true
security: []
...
paths:
...
  /v1/purchases/shippable:
    get:
      operationId: GetPurchasesShippable
      tags:
        - v1/purchases/shippable
      summary: 発送待ち購入のまとめ発送一覧の取得
      description: |
        発送可能な購入を、まとめて発送してよい組に分けて返します。認証必須で、
        管理者（admin）のみ実行できます。管理者でなければ 403 で拒否します。
        発送可能とは支払いを終え、まだ発送していない状態です。
        まとめる軸は同一の購入者で、同じ購入者宛ての購入が 1 つの組になります。購入が 1 件だけの購入者も
        その 1 件からなる組になります。
        組の中は注文日時の古い順（同時刻は購入 ID の昇順）、組同士はその組の最も古い購入の同じ順序で並びます。
        limit は読み出す購入の件数で、まとめ判定はその範囲の中で行います。範囲の外にある同一購入者の購入は
        別の便になります。ページネーションは行わず、上位 limit 件のみを対象とします。
        発送待ちの購入がない場合はエラーではなく空配列を返します。
        発送の実行は購入 1 件ずつ `PATCH /v1/purchases/{purchaseCode}/ship` で行います。
        本 op 自体は DB の SELECT のみで外部依存を持ちませんが、認証段（外部 IdP の JWKS 参照）の
        一時障害で応答不能となり得るため、認証必須 op の先例に倣い 503 を宣言します。
      security:
        - BearerAuth: []
      parameters:
        - $ref: '#/components/parameters/ShippableLimitParam'
      responses:
        '200':
          description: まとめて発送してよい組に分けた発送待ち購入の一覧
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PurchaseShippableResponse'
        '400':
          $ref: '#/components/responses/BadRequest400'
        '401':
          $ref: '#/components/responses/Unauthorized401'
        '403':
          $ref: '#/components/responses/Forbidden403'
        '405':
          $ref: '#/components/responses/MethodNotAllowed405'
        '500':
          $ref: '#/components/responses/InternalServerError500'
        '503':
          $ref: '#/components/responses/ServiceUnavailable503'
...
  /v1/purchases/{purchaseCode}/ship:
    patch:
      operationId: PatchPurchasesShip
      tags:
        - v1/purchases/detail/ship
      summary: 購入の発送
      description: |
        購入を発送済みへ遷移させます。認証必須で、管理者（admin）のみ実行できます。
        支払い済みからのみ遷移でき、status を「発送済み」へ更新し shippedAt をセットします。
        未払い相当（未処理 / 受付中 / 確認中 / 処理中）・完了・キャンセル済み・配達済みの購入は 409（不正遷移）、
        既に発送済みの購入も 409（二重発送）で拒否します。配送追跡（追跡番号 / 配送業者 / 追跡 URL）は扱いません。
        購入者本人であっても管理者でなければ 403 で拒否します。金額はすべて USD セント単位の整数です。
      security:
        - BearerAuth: []
      parameters:
        - $ref: '#/components/parameters/PurchaseCodeParam'
      responses:
        '200':
          description: 購入の発送に成功しました
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PurchaseShipResponse'
        '400':
          $ref: '#/components/responses/BadRequest400'
        '401':
          $ref: '#/components/responses/Unauthorized401'
        '403':
          $ref: '#/components/responses/Forbidden403'
        '404':
          $ref: '#/components/responses/NotFound404'
        '405':
          $ref: '#/components/responses/MethodNotAllowed405'
        '409':
          $ref: '#/components/responses/Conflict409'
        '500':
          $ref: '#/components/responses/InternalServerError500'
...
components:
  parameters:
...
    ShippableLimitParam:
      name: limit
      in: query
      description: |
        読み出す発送待ち購入の件数。注文日時の古い順で上位 limit 件を読み、まとめ判定はその範囲の中で行います。
        範囲の外にある同一購入者の購入は別の便になります。
      required: false
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
        example: 20
    PurchaseCodeParam:
      name: purchaseCode
      in: path
      required: true
      description: 購入コード（利用者へ注文番号として見せる一意の識別子）
      schema:
        type: string
        minLength: 1
        maxLength: 50
        example: 0190b0d4-7b1a-7c2e-9f3a-1b2c3d4e5f60
...
  securitySchemes:
    BearerAuth:
      type: http
      description: |
        認証にBearerトークンを使用することを示します。
        JWT の検証は RFC8725（JSON Web Token Best Current Practices）に従います。
        署名アルゴリズムは許可リストで固定し、`alg: none` と鍵の混同を拒否します。
      scheme: bearer
      bearerFormat: JWT
    BasicAuth:
      type: http
      description: 認証にBasic認証を使用することを示します。
      scheme: basic
  schemas:
    ErrorResponse:
      type: object
      description: |
        エラーレスポンスの共通スキーマ（base）。 details は返さない。details を返すエンドポイントは ErrorResponseWithDetails を参照する。
      required:
        - code
        - message
        - requestId
      properties:
        code:
          type: string
          description: 機械的に処理可能なエラーコード
          example: EXAMPLE_ERROR
        message:
          type: string
          description: 人間が読めるエラーメッセージ
          example: 定義されたエラーメッセージ
        requestId:
          type: string
          description: リクエストID
          example: abcdefghijklmnopqrstuvwxyz
...
    PurchaseStatusRef:
      type: object
      description: |
        購入に紐づくステータス（ID・業務キー・名称）。name は事前解決済みで、別途の名称解決 API 呼び出しは不要です。
        表示には name を、分岐にはドメインの業務キーである code を用います。
      required:
        - id
        - code
        - name
      properties:
        id:
          type: string
          format: uuid
          description: 購入ステータスID
          example: 0190b0d4-7b1a-7c2e-9f3a-1b2c3d4e5f60
        code:
          type: integer
          format: int64
          description: 購入ステータスの業務キー。値は到達順序を意味しません（完了より支払い済みのほうが大きい）。
          example: 7
        name:
          type: string
          description: 購入ステータス名
          example: 完了
...
    PurchaseDetailResponse:
      type: object
      description: 購入明細のレスポンス。unitPrice は購入時点の価格スナップショットです。
      required:
        - productId
        - quantity
        - unitPrice
      properties:
        productId:
          type: string
          format: uuid
          description: 商品ID
          example: 5266f905-1af4-477a-bc7c-e69729e22a2b
        quantity:
          type: integer
          format: int32
          description: 購入数量
          example: 2
        unitPrice:
          type: string
          pattern: ^\d+(\.\d+)?$
          description: |
            購入時点の単価スナップショット。サブセント精度を保持する decimal 文字列（USD ドル）で表します（例 "800.00"）。
            購入成立後に商品価格が変わっても本値は不変です。
          example: '800.00'
...
    PurchaseShippableItemResponse:
      type: object
      description: |
        まとめ発送の組に含まれる購入 1 件のスキーマ。発送指示は購入 1 件ずつ code を指定して行います。
        組に入っている時点で発送可能なので、status は返しません。明細も含みません（購入詳細 API で取得します）。
      required:
        - code
        - totalAmount
        - orderedAt
      properties:
        code:
          type: string
          description: 購入コード（利用者へ注文番号として見せる一意の識別子）。発送指示（`PATCH /v1/purchases/{purchaseCode}/ship`）に用います。
          example: 0190b0d4-7b1a-7c2e-9f3a-1b2c3d4e5f60
        totalAmount:
          type: integer
          format: int64
          description: 合計（小計 + 税額 + 送料）。USD セント単位の整数です。
          example: 176500
        orderedAt:
          type: string
          format: date-time
          description: 注文日時
          example: '2026-07-23T00:00:00Z'
    PurchaseDispatchGroupResponse:
      type: object
      description: |
        まとめて発送してよい購入の組。同一の購入者宛てであることを軸に分けます。
        組は算出結果であり永続化しないため、組そのものの識別子は持ちません。
      required:
        - userId
        - purchases
      properties:
        userId:
          type: string
          format: uuid
          description: 組に含まれる購入の購入者 ID。組の中では全件が同一です。
          example: 0190b0d4-7b1a-7c2e-9f3a-1b2c3d4e5f60
        purchases:
          type: array
          minItems: 1
          description: 組に含まれる購入。注文日時の古い順（同時刻は購入 ID の昇順）で並びます。
          items:
            $ref: '#/components/schemas/PurchaseShippableItemResponse'
    PurchaseShippableResponse:
      type: object
      description: |
        発送待ち購入のまとめ発送一覧のレスポンススキーマ。まとめて発送してよい組の配列を返します。
        組同士はその組の最も古い購入の注文日時の順（同時刻は購入 ID の昇順）で並びます。
      required:
        - groups
      properties:
        groups:
          type: array
          description: まとめ発送の組の一覧。発送待ちの購入がない場合は空配列です。
          items:
            $ref: '#/components/schemas/PurchaseDispatchGroupResponse'
...
    PurchaseShipResponse:
      type: object
      description: |
        購入発送後の購入情報レスポンススキーマ。status は ID・名称を事前解決済みで、
        別途の名称解決 API 呼び出しは不要です。shippedAt に発送日時が入ります。
        金額（subtotalAmount / taxAmount / shippingFee / totalAmount / 明細 unitPrice）は
        すべて USD セント単位の整数です。
      required:
        - code
        - userId
        - status
        - subtotalAmount
        - taxAmount
        - shippingFee
        - totalAmount
        - details
        - orderedAt
        - shippedAt
      properties:
        code:
          type: string
          description: 購入コード（利用者へ注文番号として見せる一意の識別子）
          example: 0190b0d4-7b1a-7c2e-9f3a-1b2c3d4e5f60
        userId:
          type: string
          format: uuid
          description: 購入したユーザーのID
          example: 550e8400-e29b-41d4-a716-446655440000
        status:
          $ref: '#/components/schemas/PurchaseStatusRef'
        subtotalAmount:
          type: integer
          format: int64
          description: 小計。明細（単価×数量）の合計。USD セント単位の整数です。
          example: 160000
        taxAmount:
          type: integer
          format: int64
          description: 税額（国内消費税）。USD セント単位の整数です。
          example: 16000
        shippingFee:
          type: integer
          format: int64
          description: 送料。USD セント単位の整数です。
          example: 500
        totalAmount:
          type: integer
          format: int64
          description: 合計（小計 + 税額 + 送料）。USD セント単位の整数です。
          example: 176500
        details:
          type: array
          description: 購入明細の配列
          items:
            $ref: '#/components/schemas/PurchaseDetailResponse'
        orderedAt:
          type: string
          format: date-time
          description: 注文日時
          example: '2026-07-23T00:00:00Z'
        shippedAt:
          type: string
          format: date-time
          description: 発送日時
          example: '2026-07-26T00:00:00Z'
...
  responses:
    MethodNotAllowed405:
      description: 許可されていないリクエスト方法です。
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    BadRequest400:
      description: リクエストが不正です。
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Unauthorized401:
      description: 認証が必要です。
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Forbidden403:
      description: 権限が必要です。
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    InternalServerError500:
      description: サーバーエラーが発生しました。
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    ServiceUnavailable503:
      description: サービスが一時的に利用できません。
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Conflict409:
      description: リソースが競合しています。
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...
    NotFound404:
      description: 指定されたリソースが存在しません。
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...